Move path conversion functions to conda.common.path
