+++ conda/activate.py
@@ -13,9 +13,7 @@ from __future__ import annotations
 
 import abc
 import json
-import ntpath
 import os
-import posixpath
 import re
 import sys
 from logging import getLogger
@@ -30,8 +28,6 @@ from os.path import (
     join,
 )
 from pathlib import Path
-from shutil import which
-from subprocess import run
 from textwrap import dedent
 from typing import TYPE_CHECKING
 
@@ -46,8 +42,8 @@ from .base.constants import (
 )
 from .base.context import ROOT_ENV_NAME, context, locate_prefix_by_name
 from .common.compat import FILESYSTEM_ENCODING, on_win
+from .common.path import _Cygwin, paths_equal, unix_path_to_win, win_path_to_unix
 from .common.path import path_identity as _path_identity
-from .common.path import paths_equal
 from .deprecations import deprecated
 
 if TYPE_CHECKING:
@@ -653,7 +649,7 @@ class _Activator(metaclass=abc.ABCMeta):
                 # MSYS2 /c/
                 # cygwin /cygdrive/c/
                 if re.match("^(/[A-Za-z]/|/cygdrive/[A-Za-z]/).*", prefix):
-                    path = unix_path_to_native(path, prefix)
+                    path = unix_path_to_win(path, prefix)
 
                 if isdir(path):
                     variants.append(variant)
@@ -883,223 +879,30 @@ def ensure_fs_path_encoding(value):
         return value
 
 
-class _Cygpath:
-    @classmethod
-    def nt_to_posix(cls, paths: str) -> str:
-        return cls.RE_UNIX.sub(cls.translate_unix, paths).replace(
-            ntpath.pathsep, posixpath.pathsep
-        )
+deprecated.constant(
+    "25.3",
+    "25.9",
+    "_Cygwin",
+    _Cygwin,
+    addendum="Use `conda.common.path._Cygwin` instead.",
+)
+del _Cygwin
 
-    RE_UNIX = re.compile(
-        r"""
-        (?P<drive>[A-Za-z]:)?
-        (?P<path>[\/\\]+(?:[^:*?\"<>|;]+[\/\\]*)*)
-        """,
-        flags=re.VERBOSE,
-    )
+deprecated.constant(
+    "25.3",
+    "25.9",
+    "native_path_to_unix",
+    win_path_to_unix,
+    addendum="Use `conda.common.path.win_path_to_unix` instead.",
+)
 
-    @staticmethod
-    def translate_unix(match: re.Match) -> str:
-        return "/" + (
-            ((match.group("drive") or "").lower() + match.group("path"))
-            .replace("\\", "/")
-            .replace(":", "")  # remove drive letter delimiter
-            .replace("//", "/")
-            .rstrip("/")
-        )
-
-    @classmethod
-    def posix_to_nt(cls, paths: str, prefix: str) -> str:
-        if posixpath.sep not in paths:
-            # nothing to translate
-            return paths
-
-        if posixpath.pathsep in paths:
-            return ntpath.pathsep.join(
-                cls.posix_to_nt(path, prefix) for path in paths.split(posixpath.pathsep)
-            )
-        path = paths
-
-        # Reverting a Unix path means unpicking MSYS2/Cygwin
-        # conventions -- in order!
-        # 1. drive letter forms:
-        #      /x/here/there - MSYS2
-        #      /cygdrive/x/here/there - Cygwin
-        #    transformed to X:\here\there -- note the uppercase drive letter!
-        # 2. either:
-        #    a. mount forms:
-        #         //here/there
-        #       transformed to \\here\there
-        #    b. root filesystem forms:
-        #         /here/there
-        #       transformed to {prefix}\Library\here\there
-        # 3. anything else
-
-        # continue performing substitutions until a match is found
-        path, subs = cls.RE_DRIVE.subn(cls.translation_drive, path)
-        if not subs:
-            path, subs = cls.RE_MOUNT.subn(cls.translation_mount, path)
-        if not subs:
-            path, _ = cls.RE_ROOT.subn(
-                lambda match: cls.translation_root(match, prefix), path
-            )
-
-        return re.sub(r"/+", r"\\", path)
-
-    RE_DRIVE = re.compile(
-        r"""
-        ^
-        (/cygdrive)?
-        /(?P<drive>[A-Za-z])
-        (/+(?P<path>.*)?)?
-        $
-        """,
-        flags=re.VERBOSE,
-    )
-
-    @staticmethod
-    def translation_drive(match: re.Match) -> str:
-        drive = match.group("drive").upper()
-        path = match.group("path") or ""
-        return f"{drive}:\\{path}"
-
-    RE_MOUNT = re.compile(
-        r"""
-        ^
-        //(
-            (?P<mount>[^/]+)
-            (?P<path>/+.*)?
-        )?
-        $
-        """,
-        flags=re.VERBOSE,
-    )
-
-    @staticmethod
-    def translation_mount(match: re.Match) -> str:
-        mount = match.group("mount") or ""
-        path = match.group("path") or ""
-        return f"\\\\{mount}{path}"
-
-    RE_ROOT = re.compile(
-        r"""
-        ^
-        (?P<path>/[^:]*)
-        $
-        """,
-        flags=re.VERBOSE,
-    )
-
-    @staticmethod
-    def translation_root(match: re.Match, prefix: str) -> str:
-        path = match.group("path")
-        return f"{prefix}\\Library{path}"
-
-
-def native_path_to_unix(
-    paths: str | Iterable[str] | None,
-) -> str | tuple[str, ...] | None:
-    if paths is None:
-        return None
-    elif not on_win:
-        return _path_identity(paths)
-
-    # short-circuit if we don't get any paths
-    paths = paths if isinstance(paths, str) else tuple(paths)
-    if not paths:
-        return "." if isinstance(paths, str) else ()
-
-    # on windows, uses cygpath to convert windows native paths to posix paths
-
-    # It is very easy to end up with a bash in one place and a cygpath in another due to e.g.
-    # using upstream MSYS2 bash, but with a conda env that does not have bash but does have
-    # cygpath.  When this happens, we have two different virtual POSIX machines, rooted at
-    # different points in the Windows filesystem.  We do our path conversions with one and
-    # expect the results to work with the other.  It does not.
-
-    bash = which("bash")
-    cygpath = str(Path(bash).parent / "cygpath") if bash else "cygpath"
-    joined = paths if isinstance(paths, str) else ntpath.pathsep.join(paths)
-
-    try:
-        # if present, use cygpath to convert paths since its more reliable
-        unix_path = run(
-            [cygpath, "--unix", "--path", joined],
-            text=True,
-            capture_output=True,
-            check=True,
-        ).stdout.strip()
-    except FileNotFoundError:
-        # fallback logic when cygpath is not available
-        # i.e. conda without anything else installed
-        log.warning("cygpath is not available, fallback to manual path conversion")
-
-        unix_path = _Cygpath.nt_to_posix(joined)
-    except Exception as err:
-        log.error("Unexpected cygpath error (%s)", err)
-        raise
-
-    if isinstance(paths, str):
-        return unix_path
-    elif not unix_path:
-        return ()
-    else:
-        return tuple(unix_path.split(posixpath.pathsep))
-
-
-def unix_path_to_native(
-    paths: str | Iterable[str] | None, prefix: str
-) -> str | tuple[str, ...] | None:
-    if paths is None:
-        return None
-    elif not on_win:
-        return _path_identity(paths)
-
-    # short-circuit if we don't get any paths
-    paths = paths if isinstance(paths, str) else tuple(paths)
-    if not paths:
-        return "." if isinstance(paths, str) else ()
-
-    # on windows, uses cygpath to convert posix paths to windows native paths
-
-    # It is very easy to end up with a bash in one place and a cygpath in another due to e.g.
-    # using upstream MSYS2 bash, but with a conda env that does not have bash but does have
-    # cygpath.  When this happens, we have two different virtual POSIX machines, rooted at
-    # different points in the Windows filesystem.  We do our path conversions with one and
-    # expect the results to work with the other.  It does not.
-
-    bash = which("bash")
-    cygpath = str(Path(bash).parent / "cygpath") if bash else "cygpath"
-    joined = paths if isinstance(paths, str) else posixpath.pathsep.join(paths)
-
-    try:
-        # if present, use cygpath to convert paths since its more reliable
-        win_path = run(
-            [cygpath, "--windows", "--path", joined],
-            text=True,
-            capture_output=True,
-            check=True,
-        ).stdout.strip()
-    except FileNotFoundError:
-        # fallback logic when cygpath is not available
-        # i.e. conda without anything else installed
-        log.warning("cygpath is not available, fallback to manual path conversion")
-
-        # The conda prefix can be in a drive letter form
-        prefix = _Cygpath.posix_to_nt(prefix, prefix)
-
-        win_path = _Cygpath.posix_to_nt(joined, prefix)
-    except Exception as err:
-        log.error("Unexpected cygpath error (%s)", err)
-        raise
-
-    if isinstance(paths, str):
-        return win_path
-    elif not win_path:
-        return ()
-    else:
-        return tuple(win_path.split(ntpath.pathsep))
-
+deprecated.constant(
+    "25.3",
+    "25.9",
+    "unix_path_to_native",
+    unix_path_to_win,
+    addendum="Use `conda.common.path.unix_path_to_win` instead.",
+)
 
 deprecated.constant(
     "25.3",
@@ -1124,7 +927,7 @@ def backslash_to_forwardslash(
 class PosixActivator(_Activator):
     pathsep_join = ":".join
     sep = "/"
-    path_conversion = staticmethod(native_path_to_unix)
+    path_conversion = staticmethod(win_path_to_unix)
     script_extension = ".sh"
     tempfile_extension = None  # output to stdout
     command_join = "\n"
@@ -1177,7 +980,7 @@ class PosixActivator(_Activator):
 class CshActivator(_Activator):
     pathsep_join = ":".join
     sep = "/"
-    path_conversion = staticmethod(native_path_to_unix)
+    path_conversion = staticmethod(win_path_to_unix)
     script_extension = ".csh"
     tempfile_extension = None  # output to stdout
     command_join = ";\n"
@@ -1280,7 +1083,7 @@ class CmdExeActivator(_Activator):
 class FishActivator(_Activator):
     pathsep_join = '" "'.join
     sep = "/"
-    path_conversion = staticmethod(native_path_to_unix)
+    path_conversion = staticmethod(win_path_to_unix)
     script_extension = ".fish"
     tempfile_extension = None  # output to stdout
     command_join = ";\n"
